pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lentel"
version = "1.0.0"
description = "A new UDP-based protocol for fast, NAT-free file transfer of any size."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [{ name = "Lentel contributors" }]
keywords = ["file-transfer", "udp", "nat-traversal", "p2p", "protocol"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet",
    "Topic :: System :: Networking",
]
dependencies = [
    "cryptography>=41.0",
]

[project.optional-dependencies]
tray = [
    "pystray>=0.19",
    "Pillow>=10.0",
]

[project.scripts]
lentel = "lentel.cli:main"
lentel-coordinator = "lentel.server.coordinator:main"
lentel-relay = "lentel.server.relay:main"

[project.gui-scripts]
# Registered as a gui-script so Windows doesn't open a console window.
lentel-tray = "lentel.app.__main__:main"

[tool.setuptools.packages.find]
include = ["lentel*"]
